requirements.txt
flask==3.0.0
pyyaml==6.0.1
pytest==7.4.0